arm64: add alternative runtime patching
With a blatant copy of some x86 bits we introduce the alternative
runtime patching "framework" to arm64.
This is quite basic for now and we only provide the functions we need
at this time.
This is connected to the newly introduced feature bits.

Change-Id: Ib52ad3aaaa8cfd9e6dafdd78e7386ae2553e9b68
Signed-off-by: Andre Przywara <andre.przywara@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>
Reviewed-on: http://git-master/r/721972
(cherry picked from commit 1535a1abdf4584adca4dcc0b3b0a8071e6160001)
Reviewed-on: http://git-master/r/731446
Reviewed-by: Bo Yan <byan@nvidia.com>
Tested-by: Bo Yan <byan@nvidia.com>