drivers: tegra: dc: fix tegra_dc_probe failure
in case tegra_dc_probe() is failed, then we should
mark the corresponding index entry in array
tegra_dcs[] as NULL.

Since this array will hold the pointer to tegra_dc
struct which is freed in failure case. Therefore,
if in future any driver try to access this array
that driver will find the pointer which is
already freed.

bug 1629420

Change-Id: I4428d7104da2272ac0827e9f169ebb959beefdd1
Signed-off-by: Pankaj Kumar <pankumar@nvidia.com>
Reviewed-on: http://git-master/r/764810
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>